Fix concept-label dtype mismatch in CUB_DatasetGenerator caching
Uncached __getitem__ returned attribute_label as int64, while the caching
path (_unpack_attributes) returned float32. Once enough training indices
get cached across epochs, a batch can mix both, and PyTorch's collate
crashes trying to stack tensors of different dtypes into a reused buffer.

+++ datasets/CUB_dataset.py
@@ -214,7 +214,7 @@ class CUB_DatasetGenerator(Dataset):
             img_path = img_data["img_path"] 
             image_data = Image.open(img_path).convert("RGB") 
             image_label = img_data["class_label"] 
-            image_attr = np.array(img_data["attribute_label"])
+            image_attr = np.array(img_data["attribute_label"], dtype=np.float32)
             
             if self.cache:
                 self._cache_image(index, image_data, image_attr, image_label)
